+++ src/main/java/carrotbat410/lol/jwt/JWTUtil.java
@@ -29,8 +29,8 @@ public class JWTUtil {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("role", String.class);
     }
 
-    public Integer getId(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("id", Integer.class);
+    public Long getId(String token) {
+        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("id", Long.class);
     }
 
     public Boolean isExpired(String token) {
@@ -38,7 +38,7 @@ public class JWTUtil {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().getExpiration().before(new Date());
     }
 
-    public String createJwt(String username, String role, Integer id, Long expiredMs) {
+    public String createJwt(String username, String role, Long id, Long expiredMs) {
 
         return Jwts.builder()
                 .claim("username", username)
